application/nextjs-app-v4/application/nextjs-app-fetch/backend/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
environ==1.0
mysqlclient==2.2.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
